// src/interfaces/videos.interface.ts
export interface VideoResponseInterface {
  id: string;
  name: string;
  url_240p: string;
  url_480p: string;
  url_1080p: string;
  url_4k: string;
}

export interface VideoInterface {
  id: number;
  public_id: string;
  name: string;
  url_240p: string;
  url_480p: string;
  url_1080p: string;
  url_4k: string;
}